PIC-387: Refactor - move convictions to OffenderController
